pyproject.toml
[tool.poetry]
name = "mangostar"
version = "2.6.0"
description = ""
authors = ["karamel <kivo360@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
SQLAlchemy = "^1.3.20"
psycopg2 = "^2.8.6"
pydantic = {extras = ["email"], version = "^1.8.1"}
pampy = "^0.3.0"
loguru = "^0.5.3"
auto-all = "^1.3.0"
typer = "^0.3.2"
yapf = "^0.30.0"
addict = "^2.4.0"
networkx = "^2.5"
genson = "^1.2.2"
inflection = "^0.5.1"
devtools = "^0.6.1"
orjson = "^3.4.8"
frozendict = "^1.2"
stringcase = "^1.2.0"
Faker = "^6.4.1"
more-itertools = "^8.7.0"
toolz = "^0.11.1"
cytoolz = "^0.11.0"
sqlalchemy-plus = "^0.2.0"
decorator = "^4.4.2"
Random-Word = "^1.0.7"
PyYAML = "^5.4.1"
pglast = "^1.17"
Mako = "^1.1.4"
humanize = "^3.2.0"
maya = "^0.6.1"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
httpx = "^0.17.0"
python-arango = "^7.1.0"
xxhash = "^2.0.0"
pyrsistent = "^0.17.3"
respx = "^0.16.3"
retry = "^0.9.2"
graphviz = "^0.19.1"
pygraphviz = "^1.9"
transitions = {extras = ["diagrams"], version = "^0.8.11"}
alembic = "^1.7.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
hypothesis = {extras = ["dateutil", "cli"], version = "^6.4.3"}
newsapi-python = "^0.2.6"
numpy = "^1.20.1"
pandas = "^1.2.3"
yapf = "^0.30.0"
flake8 = "^3.8.4"
mypy = "^0.800"
isort = "^5.7.0"
pytest-sugar = "^0.9.4"
faker-schema = "^0.1.4"
vulture = "^2.3"
black = {version = "^20.8b1", allow-prereleases = true}
jupyterlab = "^3.0.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "google"
lines_after_imports=2
force_sort_within_sections=false
known_devtools=['devtools','loguru', 'inflection']
sections=['FUTURE','STDLIB', 'DEVTOOLS','THIRDPARTY','FIRSTPARTY','LOCALFOLDER']


[tool.pytest.ini_options]
addopts = "-ra -q -s --capture=no"

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}


[tool.vulture]
exclude = ["service/", "imports.py"]
ignore_decorators = ["@app.*", "@require_*"]
ignore_names = ["test_*"]
make_whitelist = true
min_confidence = 80
# paths = ["myscript.py", "mydir"]
sort_by_size = true
# verbose = true